Fix building Docker image with Nix
I was trying to change the "system" attribute of the derivation, after
Nixpkgs had already done everything under the assumption I was building
for "x86_64-darwin." I needed to configure the Nixpkgs function with the
"system" attribute so that it propagates to everything else.

I also managed to clean things up so that my overrides don't clobber
each other.