Simplify Haskell packages overlay
Previously, in order to override the Haskell package set, you had to
generate a derivation using `cabal2nix`, and then add it to the overlay
manually. This change removes that second step. Now all Haskell package
derivations under `nix/haskell-packages/` are read in automatically.

This also means that any changes to the Haskell derivation, such as
turning off tests, must be done in the derivation, and not through an
`overrideCabal` helper function in the overlay config. I'm not sure how
I feel about this yet -- I prefer to leave output from `cabal2nix` (or
any other code-gen tool) untouched -- however I must admit it does
simplify things.

Finally, this also means that this project's derivation gets its own
derivation under `nix/haskell-packages/`. I like this because it treats
`converge` like any other Haskell package in the package set. It doesn't
matter whether it's from Hackage, GitHub, or the current project.

The only difference is that the `converge` derivation uses
import-from-derivation with `callCabal2nix` in order to stay updated as
the code in this repo changes.